requirements.txt
python-docx
streamlit
pandas